feat: split publish pipeline into modules with retry/backoff and checkpoint/resume
- Split monolithic publish_to_hf.py (443 lines) into pipeline/ package:
  - fetch.py: data collection with exponential backoff on 429/5xx
  - transform.py: Polars transformation + schema validation
  - package.py: Parquet writing + dataset card generation
  - publish.py: HuggingFace Hub upload
- Add --resume flag to continue from last successful fetch
- Checkpoint saved after each param set fetch for crash recovery
- Existing CLI interface preserved (backward compatible)

Closes #59
