+++ scripts/pipeline/__init__.py
@@ -1,0 +1,15 @@
+"""Pipeline modules for kpubdata HuggingFace publishing."""
+
+from pipeline.fetch import fetch_records
+from pipeline.package import generate_dataset_card, write_parquet
+from pipeline.publish import upload_to_hf
+from pipeline.transform import transform_records, validate_schema
+
+__all__ = [
+    "fetch_records",
+    "transform_records",
+    "validate_schema",
+    "write_parquet",
+    "generate_dataset_card",
+    "upload_to_hf",
+]
